refactor: platforms object
