+++ lib/Version.js
@@ -1,4 +1,4 @@
-var platforms = require("./platforms");
+var { platforms } = require("../src/utilities");
 var _ = require("lodash");
 var semver = require("semver");
