+++ lib/index.js
@@ -12,13 +12,13 @@ var semver = require("semver");
 var platformOverrides = require("./platformOverrides.js");
 var deprecate = require("deprecate");
 var updateNotifier = require("update-notifier");
-var { checkCache, detectCurrentPlatform } = require('../src/utilities');
 
 var NwVersions = require("./versions");
 var Version = require("./Version");
 var Utils = require("./utils");
 var Downloader = require("./downloader");
-var platforms = require("./platforms");
+
+var { checkCache, detectCurrentPlatform, platforms } = require('../src/utilities');
 
 var pkg = require("../package.json");
