+++ lib/versions.js
@@ -1,4 +1,4 @@
-var platforms = require("./platforms");
+var { platforms } = require("../src/utilities");
 var semver = require("semver");
 var request = require("request");
 var _ = require("lodash");
