+++ src/utilities/index.js
@@ -1,7 +1,9 @@
 const checkCache = require("./checkCache");
 const detectCurrentPlatform = require("./detectCurrentPlatform");
+const platforms = require("./platforms");
 
 module.exports = {
     checkCache,
     detectCurrentPlatform,
+    platforms,
 };
